pitext_desktop/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
anthropic==0.18.1
python-dotenv==1.0.0
playwright==1.40.0
pillow==10.1.0
python-multipart==0.0.6
httpx==0.25.2
Flask>=2.3.0
asgiref>=3.0.0
